% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_roundabout.R
\name{calc_path}
\alias{calc_path}
\title{path calculation}
\usage{
calc_path(
  island_radius,
  approach_radius,
  road_width,
  veh_width,
  buffer,
  splitter_width
)
}
\description{
path calculation
}
