% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoundaboutChallenge-package.R
\docType{package}
\name{RoundaboutChallenge-package}
\alias{RoundaboutChallenge}
\alias{RoundaboutChallenge-package}
\title{RoundaboutChallenge: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
